Update Kmeans_clustering.ipynb
K-means clustering is a method of vector quantization, originally from signal processing, that aims to partition n observations into k clusters in which each observation belongs to the cluster with the nearest mean, serving as a prototype of the cluster. The goal of kmeans is to group data points into distinct non-overlapping subgroups. K-means clustering is one of the most popular clustering algorithms. Although there is one disadvantage that we have to predict the value of "k".