Update Gradient_Descent.ipynb
Linear regression is a linear approach to modelling the relationship between a dependent variable and one or more independent variables. Gradient descent is an iterative optimization algorithm to find the minimum of a function. Our objective while using gradient descent is to minimize the cost function. Gradient descent is one of the simplest and widely used algorithms in machine learning, mainly because it can be applied to any function to optimize it.