Update decision_tree.ipynb
A decision tree is a diagram or chart that people use to determine a course of action or show a statistical probability. A decision tree is drawn upside down with its root at the top. It forms the outline of the namesake woody plant, usually upright but sometimes lying on its side. Each branch of the decision tree represents a possible decision, outcome, or reaction. The farthest branches on the tree represent the end results. Decision trees are used for handling non-linear data sets effectively.They provide a way to present algorithms with conditional control statements.